+++ packages/headless/src/api/search/search-api-client.test.ts
@@ -25,6 +25,12 @@ import {buildMockSearchResponse} from '../../test/mock-search-response';
 import {createMockState} from '../../test/mock-state';
 import {PlatformClient, PlatformClientCallOptions} from '../platform-client';
 import {NoopPreprocessRequest} from '../preprocess-request';
+import {FacetSearchRequest} from './facet-search/facet-search-request';
+import {HtmlRequest} from './html/html-request';
+import {PlanRequest} from './plan/plan-request';
+import {ProductRecommendationsRequest} from './product-recommendations/product-recommendations-request';
+import {QuerySuggestRequest} from './query-suggest/query-suggest-request';
+import {RecommendationRequest} from './recommendation/recommendation-request';
 import {
   isErrorResponse,
   SearchAPIClient,
@@ -35,6 +41,7 @@ import {
   getOrganizationIdQueryParam,
 } from './search-api-params';
 import {QuestionsAnswers} from './search/question-answering';
+import {SearchRequest} from './search/search-request';
 import {SearchResponseSuccess} from './search/search-response';
 
 jest.mock('../platform-client');
@@ -172,7 +179,9 @@ describe('search api client', () => {
       searchAPIClient.search(req);
       const request = (PlatformClient.call as jest.Mock).mock.calls[0][0];
 
-      const expectedRequest: PlatformClientCallOptions = {
+      const expectedRequest: PlatformClientCallOptions<
+        Omit<SearchRequest, 'accessToken' | 'organizationId' | 'url'>
+      > = {
         accessToken: state.configuration.accessToken,
         method: 'POST',
         contentType: 'application/json',
@@ -182,6 +191,8 @@ describe('search api client', () => {
         logger,
         origin: 'searchApiFetch',
         requestParams: {
+          referrer: state.configuration.analytics.originLevel3,
+          tab: state.configuration.analytics.originLevel2,
           q: state.query.q,
           debug: false,
           locale: state.configuration.search.locale,
@@ -257,7 +268,9 @@ describe('search api client', () => {
       searchAPIClient.plan(req);
       const request = (PlatformClient.call as jest.Mock).mock.calls[0][0];
 
-      const expectedRequest: PlatformClientCallOptions = {
+      const expectedRequest: PlatformClientCallOptions<
+        Omit<PlanRequest, 'accessToken' | 'organizationId' | 'url'>
+      > = {
         accessToken: state.configuration.accessToken,
         method: 'POST',
         contentType: 'application/json',
@@ -307,7 +320,9 @@ describe('search api client', () => {
       searchAPIClient.querySuggest(req);
       const request = (PlatformClient.call as jest.Mock).mock.calls[0][0];
 
-      const expectedRequest: PlatformClientCallOptions = {
+      const expectedRequest: PlatformClientCallOptions<
+        Omit<QuerySuggestRequest, 'accessToken' | 'organizationId' | 'url'>
+      > = {
         accessToken: state.configuration.accessToken,
         method: 'POST',
         contentType: 'application/json',
@@ -322,6 +337,7 @@ describe('search api client', () => {
           context: state.context.contextValues,
           pipeline: state.pipeline,
           searchHub: state.searchHub,
+          tab: state.configuration.analytics.originLevel2,
           timezone: state.configuration.search.timezone,
           locale: state.configuration.search.locale,
           actionsHistory: expect.any(Array),
@@ -367,7 +383,9 @@ describe('search api client', () => {
         searchAPIClient.facetSearch(req);
 
         const request = (PlatformClient.call as jest.Mock).mock.calls[0][0];
-        const expectedRequest: Partial<PlatformClientCallOptions> = {
+        const expectedRequest: PlatformClientCallOptions<
+          Omit<FacetSearchRequest, 'accessToken' | 'organizationId' | 'url'>
+        > = {
           accessToken: state.configuration.accessToken,
           method: 'POST',
           contentType: 'application/json',
@@ -376,6 +394,17 @@ describe('search api client', () => {
           }/facet?${getOrganizationIdQueryParam(req)}`,
           requestMetadata: {
             method: 'facetSearch',
+          },
+          logger,
+          origin: 'searchApiFetch',
+          preprocessRequest: NoopPreprocessRequest,
+          requestParams: {
+            captions: expect.any(Object),
+            field: facetState.request.field,
+            filterFacetCount: req.filterFacetCount,
+            numberOfValues: facetSearchState.options.numberOfValues,
+            query: `*${facetSearchState.options.query}*`,
+            type: facetState.request.type,
           },
         };
 
@@ -445,7 +474,7 @@ describe('search api client', () => {
       });
 
       it(`when the id is on the categoryFacetSearchSet,
-      it calls PlatformClient.call with the category facet search params`, async () => {
+it calls PlatformClient.call with the category facet search params`, async () => {
         const id = '1';
         const categoryFacetSearch = buildMockCategoryFacetSearch();
         const categoryFacet = buildMockCategoryFacetRequest();
@@ -511,7 +540,9 @@ describe('search api client', () => {
 
       searchAPIClient.recommendations(req);
 
-      const expectedRequest: PlatformClientCallOptions = {
+      const expectedRequest: PlatformClientCallOptions<
+        Omit<RecommendationRequest, 'accessToken' | 'organizationId' | 'url'>
+      > = {
         accessToken: recommendationState.configuration.accessToken,
         method: 'POST',
         contentType: 'application/json',
@@ -526,7 +557,8 @@ describe('search api client', () => {
           cq: recommendationState.advancedSearchQueries.cq,
           fieldsToInclude: recommendationState.fields.fieldsToInclude,
           context: recommendationState.context.contextValues,
-          dictionaryFieldContext: recommendationState.context.contextValues,
+          dictionaryFieldContext:
+            recommendationState.dictionaryFieldContext.contextValues,
           pipeline: recommendationState.pipeline,
           searchHub: recommendationState.searchHub,
           timezone: state.configuration.search.timezone,
@@ -567,7 +599,7 @@ describe('search api client', () => {
     });
 
     it(`when calling SearchAPIClient.productRecommendations
-      should call PlatformClient.call with the right options`, async () => {
+should call PlatformClient.call with the right options`, async () => {
       const productRecommendationsState = buildMockProductRecommendationsState({
         productRecommendations: {
           ...getProductRecommendationsInitialState(),
@@ -590,7 +622,12 @@ describe('search api client', () => {
       searchAPIClient.productRecommendations(req);
       const request = (PlatformClient.call as jest.Mock).mock.calls[0][0];
 
-      const expectedRequest: PlatformClientCallOptions = {
+      const expectedRequest: PlatformClientCallOptions<
+        Omit<
+          ProductRecommendationsRequest,
+          'accessToken' | 'organizationId' | 'url'
+        >
+      > = {
         accessToken: productRecommendationsState.configuration.accessToken,
         method: 'POST',
         contentType: 'application/json',
@@ -629,13 +666,19 @@ describe('search api client', () => {
     });
 
     it(`when calling SearchAPIClient.fieldDescriptions
-    should call PlatformClient.call with the right options`, async () => {
+should call PlatformClient.call with the right options`, async () => {
       const req = (await buildSearchRequest(state)).request;
 
       searchAPIClient.fieldDescriptions(req);
       const request = (PlatformClient.call as jest.Mock).mock.calls[0][0];
 
-      const expectedRequest: Partial<PlatformClientCallOptions> = {
+      const expectedRequest: PlatformClientCallOptions<{}> = {
+        accessToken: state.configuration.accessToken,
+        contentType: 'application/json',
+        logger,
+        origin: 'searchApiFetch',
+        preprocessRequest: NoopPreprocessRequest,
+        requestParams: {},
         requestMetadata: {method: 'fieldDescriptions'},
         method: 'GET',
         url: `${
@@ -674,7 +717,24 @@ describe('search api client', () => {
         const req = await buildResultPreviewRequest(state, {uniqueId: '1'});
         searchAPIClient.html(req);
         const request = (PlatformClient.call as jest.Mock).mock.calls[0][0];
-        const expectedRequest: Partial<PlatformClientCallOptions> = {
+        const expectedRequest: PlatformClientCallOptions<
+          Omit<HtmlRequest, 'accessToken' | 'organizationId' | 'url'>
+        > = {
+          accessToken: state.configuration.accessToken,
+          contentType: 'application/x-www-form-urlencoded',
+          logger,
+          method: 'POST',
+          origin: 'searchApiFetch',
+          url: `${
+            state.configuration.search.apiBaseUrl
+          }/html?${getOrganizationIdQueryParam(req)}`,
+          preprocessRequest: NoopPreprocessRequest,
+          requestParams: {
+            enableNavigation: false,
+            q: state.query.q,
+            requestedOutputSize: req.requestedOutputSize,
+            uniqueId: req.uniqueId,
+          },
           requestMetadata: {method: 'html'},
         };
         expect(request).toMatchObject(expectedRequest);