+++ packages/headless/src/controllers/commerce/search/pagination/headless-search-pagination.ts
@@ -4,6 +4,7 @@ import {loadReducerError} from '../../../../utils/errors';
 import {
   buildCorePagination,
   Pagination,
+  PaginationProps,
 } from '../../core/pagination/headless-core-commerce-pagination';
 import {loadSearchReducer} from '../utils/load-search-reducers';
 
@@ -11,14 +12,19 @@ import {loadSearchReducer} from '../utils/load-search-reducers';
  * Creates a `Pagination` controller instance.
  *
  * @param engine - The headless commerce engine.
+ * @param props - The configurable `Pagination` controller properties.
  * @returns A `Pagination` controller instance.
  * */
-export function buildSearchPagination(engine: CommerceEngine): Pagination {
+export function buildSearchPagination(
+  engine: CommerceEngine,
+  props: PaginationProps = {}
+): Pagination {
   if (!loadSearchReducer(engine)) {
     throw loadReducerError;
   }
 
   return buildCorePagination(engine, {
+    ...props,
     fetchResultsActionCreator: executeSearch,
   });
 }