+++ packages/headless/src/features/commerce/pagination/pagination-slice.test.ts
@@ -12,7 +12,12 @@ import {
 import {setContext, setUser, setView} from '../context/context-actions';
 import {fetchProductListing} from '../product-listing/product-listing-actions';
 import {executeSearch} from '../search/search-actions';
-import {nextPage, previousPage, selectPage} from './pagination-actions';
+import {
+  nextPage,
+  previousPage,
+  selectPage,
+  setPageSize,
+} from './pagination-actions';
 import {paginationReducer} from './pagination-slice';
 import {
   CommercePaginationState,
@@ -36,7 +41,7 @@ describe('pagination slice', () => {
     const finalState = paginationReducer(undefined, {type: ''});
     expect(finalState).toEqual({
       page: 0,
-      perPage: 0,
+      perPage: undefined,
       totalCount: 0,
       totalPages: 0,
     });
@@ -89,6 +94,13 @@ describe('pagination slice', () => {
     const finalState = paginationReducer(state, previousPage());
 
     expect(finalState.page).toBe(0);
+  });
+
+  it('#setPageSize sets the page size', () => {
+    const pageSize = 17;
+    const finalState = paginationReducer(state, setPageSize(pageSize));
+
+    expect(finalState.perPage).toBe(pageSize);
   });
 
   it('sets the pagination on #fetchProductListing.fulfilled', () => {
@@ -147,11 +159,13 @@ describe('pagination slice', () => {
   ])('$actionName', ({action}) => {
     it('resets pagination', () => {
       state.page = 5;
+      state.perPage = 17;
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       const finalState = paginationReducer(state, action({} as any));
 
       expect(finalState.page).toBe(0);
+      expect(finalState.perPage).toBe(undefined);
     });
   });
 });