+++ packages/headless/src/state/commerce-app-state.ts
@@ -4,6 +4,7 @@ import {
   CommerceContextSection,
   ConfigurationSection,
   ProductListingV2Section,
+  RecommendationsSection,
   VersionSection,
   CommerceSortSection,
   CommerceSearchSection,
@@ -23,6 +24,7 @@ export type CommerceProductListingParametersState = {};
 
 export type CommerceAppState = ConfigurationSection &
   ProductListingV2Section &
+  RecommendationsSection &
   CommerceSearchSection &
   CommerceQuerySection &
   FacetOrderSection &